Move provider endpoint tests into separate files
