+++ packages/api/tests/endpoints/providers/update-provider.js
@@ -1,0 +1,100 @@
+'use strict';
+
+const test = require('ava');
+const { randomString } = require('@cumulus/common/test-utils');
+
+const bootstrap = require('../../../lambdas/bootstrap');
+const models = require('../../../models');
+const providerEndpoint = require('../../../endpoints/providers');
+const {
+  fakeUserFactory,
+  fakeProviderFactory,
+  testEndpoint
+} = require('../../../lib/testUtils');
+const { Search } = require('../../../es/search');
+const assertions = require('../../../lib/assertions');
+
+process.env.UsersTable = randomString();
+process.env.ProvidersTable = randomString();
+process.env.stackName = randomString();
+process.env.internal = randomString();
+let providerModel;
+const esIndex = randomString();
+let esClient;
+
+let authHeaders;
+let userModel;
+
+test.before(async () => {
+  await bootstrap.bootstrapElasticSearch('fakehost', esIndex);
+
+  providerModel = new models.Provider();
+  await providerModel.createTable();
+
+  userModel = new models.User();
+  await userModel.createTable();
+
+  const authToken = (await userModel.create(fakeUserFactory())).password;
+  authHeaders = {
+    Authorization: `Bearer ${authToken}`
+  };
+
+  esClient = await Search.es('fakehost');
+});
+
+test.beforeEach(async (t) => {
+  t.context.testProvider = fakeProviderFactory();
+  await providerModel.create(t.context.testProvider);
+});
+
+test.after.always(async () => {
+  await providerModel.deleteTable();
+  await userModel.deleteTable();
+  await esClient.indices.delete({ index: esIndex });
+});
+
+test('CUMULUS-911 PUT with pathParameters and without an Authorization header returns an Authorization Missing response', async (t) => {
+  const request = {
+    httpMethod: 'PUT',
+    pathParameters: {
+      id: 'asdf'
+    },
+    headers: {}
+  };
+
+  return testEndpoint(providerEndpoint, request, (response) => {
+    assertions.isAuthorizationMissingResponse(t, response);
+  });
+});
+
+test('CUMULUS-912 PUT with pathParameters and with an unauthorized user returns an unauthorized response', async (t) => {
+  const request = {
+    httpMethod: 'PUT',
+    pathParameters: {
+      id: 'asdf'
+    },
+    headers: {
+      Authorization: 'Bearer invalid-token'
+    }
+  };
+
+  return testEndpoint(providerEndpoint, request, (response) => {
+    assertions.isUnauthorizedUserResponse(t, response);
+  });
+});
+
+test('PUT updates an existing provider', (t) => {
+  const updatedLimit = 2;
+
+  const putEvent = {
+    httpMethod: 'PUT',
+    pathParameters: { id: t.context.testProvider.id },
+    body: JSON.stringify({ globalConnectionLimit: updatedLimit }),
+    headers: authHeaders
+  };
+
+  return testEndpoint(providerEndpoint, putEvent, (response) => {
+    const { globalConnectionLimit } = JSON.parse(response.body);
+    t.is(globalConnectionLimit, updatedLimit);
+  });
+});
